requirements.txt
# --- Desarrollo e Interacción ---
ipython==8.37.0
jupyterlab==4.4.10
notebook==7.4.7

# --- Kedro y plugins ---
click==8.1.8
kedro==1.0.0
kedro-datasets==8.1.0
kedro-telemetry==0.6.5
kedro-viz==12.1.0
pyspark>=3.5,<4.0

# --- Machine Learning y análisis ---
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.5.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# --- Airflow y Providers (versión controlada por constraints.txt) ---
apache-airflow==3.1.2rc1
apache-airflow-providers-docker==3.7.0
apache-airflow-providers-postgres==5.5.0
apache-airflow-providers-ssh==3.6.0

# --- Otros ---
docker==6.1.3
psycopg2-binary==2.9.7
setuptools>=68.0.0
wheel
